🐛 [fix] 리뷰 반영해 수정
