+++ Projects/LastDance/LastDance/Sources/Domain/Model/ClipPayload.swift
@@ -1,0 +1,14 @@
+//
+//  ClipPayload.swift
+//  LastDance
+//
+//  Created by 배현진 on 11/13/25.
+//
+
+struct ClipPayload {
+    let visitorUUID: String
+    let visitorId: Int
+    let visitId: Int
+    let exhibitionId: Int
+    let artworkId: Int?
+}
