+++ Projects/LastDance/LastDanceClip/Sources/View/ClipArtReactionView.swift
@@ -30,7 +30,7 @@ struct ClipArtReactionView: View {
     
     var body: some View {
         // 스크롤에 따라 이미지 크기 조정
-        let imageHeight = max(200, 388 - scrollOffset * 0.5)
+        let imageHeight = max(100, 388 - scrollOffset * 1.5)
         let imageWidth = max(150, 281 - (388 - imageHeight) * 0.76)
         
         ZStack(alignment: .top) {
@@ -42,26 +42,23 @@ struct ClipArtReactionView: View {
                         Color.clear.frame(height: 36)
                         
                         // 작품 이미지
-                        if let imageURLString = viewModel.artwork?.thumbnailURL,
-                           let url = URL(string: imageURLString) {
-                            AsyncImage(url: url) { image in
-                                image
-                                    .resizable()
-                                    .aspectRatio(contentMode: .fill)
-                            } placeholder: {
-                                ProgressView()
-                            }
-                            .frame(width: imageWidth, height: isMessageFieldFocused ? 0 : imageHeight)
-                            .cornerRadius(24)
+                        if let imageURLString = viewModel.artwork?.thumbnailURL {
+                            CachedImage(
+                                imageURLString,
+                                targetSize: .init(width: imageWidth, height: imageHeight)
+                            )
+                            .aspectRatio(contentMode: .fill)
+                            .frame(width: imageWidth, height: imageHeight)
+                            .clipShape(RoundedRectangle(cornerRadius: 24))
                             .overlay(
                                 RoundedRectangle(cornerRadius: 24)
-                                    .stroke(Color.color5, lineWidth: 10)
+                                    .stroke(LDColor.color5, lineWidth: 10)
                             )
                             .shadow(color: Color.black.opacity(0.25), radius: 2, x: 0, y: 0)
                         } else {
                             Rectangle()
                                 .fill(Color.gray.opacity(0.15))
-                                .frame(width: imageWidth, height: isMessageFieldFocused ? 0 : imageHeight)
+                                .frame(width: imageWidth, height: imageHeight)
                                 .cornerRadius(24)
                                 .overlay(Text("이미지 없음").foregroundColor(.gray))
                         }
@@ -79,7 +76,7 @@ struct ClipArtReactionView: View {
                         }
                     }
                 }
-                .background(Color.white)
+                .background(LDColor.color6)
                 
             } else {
                 // 처음 로딩 상태
@@ -88,17 +85,17 @@ struct ClipArtReactionView: View {
                     ProgressView()
                     Spacer()
                 }
-                .background(Color.white)
+                .background(LDColor.color6)
             }
             
             // 최상단 고정 탭바
             VStack(spacing: 0) {
                 Color.clear
                     .frame(height: 36)
-                    .background(Color.white)
+                    .background(LDColor.color6)
                 
                 TabBarView(selectedTab: $selectedTab)
-                    .background(Color.white)
+                    .background(LDColor.color6)
             }
             // 고정 탭바 표시 조건: 스크롤 임계값 도달 OR 메시지 필드에 포커스
             .opacity((viewModel.isTabBarFixed(for: scrollOffset) || isMessageFieldFocused) ? 1 : 0)
@@ -113,7 +110,7 @@ struct ClipArtReactionView: View {
                         isEnabled: viewModel.hasText,
                         action: { buttonAction() }
                     )
-                    .background(Color.white)
+                    .background(LDColor.color6)
                 }
             }
         }
@@ -124,8 +121,13 @@ struct ClipArtReactionView: View {
     }
     
     private func buttonAction() {
-        viewModel.sendReaction()
-        router.push(.complete)
+        Task { @MainActor in
+            if await viewModel.sendReaction() {
+                router.push(.complete)
+            } else {
+                // TODO: - 전송 실패 알럿 띄우기 (버튼 throttle 함께 처리)
+            }
+        }
     }
     
     // MARK: - 작품 탭
@@ -133,13 +135,13 @@ struct ClipArtReactionView: View {
         VStack(alignment: .leading, spacing: 12) {
             Text(viewModel.artwork?.title ?? "작품명")
                 .font(.system(size: 21, weight: .semibold))
-                .foregroundColor(Color.black)
+                .foregroundColor(LDColor.color1)
             
             if let artistName = viewModel.artistName {
                 HStack {
                     Text(artistName)
                         .font(.system(size: 16, weight: .medium))
-                        .foregroundColor(Color.color1)
+                        .foregroundColor(LDColor.color1)
                         .padding(.horizontal, 10)
                         .padding(.vertical, 4)
                         .background(
@@ -147,7 +149,7 @@ struct ClipArtReactionView: View {
                                 .stroke(Color.black, lineWidth: 1)
                                 .background(
                                     RoundedRectangle(cornerRadius: 24)
-                                        .fill(Color.white)
+                                        .fill(LDColor.color6)
                                 )
                         )
                     Spacer()
@@ -156,7 +158,7 @@ struct ClipArtReactionView: View {
                 HStack {
                     Text("작자미상")
                         .font(.system(size: 16, weight: .medium))
-                        .foregroundColor(Color.color1)
+                        .foregroundColor(LDColor.color1)
                         .padding(.horizontal, 10)
                         .padding(.vertical, 4)
                         .background(
@@ -164,7 +166,7 @@ struct ClipArtReactionView: View {
                                 .stroke(Color.black, lineWidth: 1)
                                 .background(
                                     RoundedRectangle(cornerRadius: 24)
-                                        .fill(Color.white)
+                                        .fill(LDColor.color6)
                                 )
                         )
                     Spacer()
@@ -184,7 +186,7 @@ struct ClipArtReactionView: View {
                 
                 Text(desc)
                     .font(.system(size: 16, weight: .medium))
-                    .foregroundColor(Color.black.opacity(0.8))
+                    .foregroundColor(LDColor.color2)
                     .lineSpacing(4)
                     .fixedSize(horizontal: false, vertical: true)
             } else {
@@ -215,6 +217,7 @@ struct ClipArtReactionView: View {
                 
                 if viewModel.message.isEmpty && !isMessageFieldFocused {
                     Text("작품에 대한 생각을 자유롭게 적어보세요.")
+                        .font(.system(size: 16))
                         .foregroundColor(LDColor.color3)
                         .padding(.top, 12)
                         .padding(.leading, 14)
@@ -236,7 +239,7 @@ struct ClipArtReactionView: View {
                     HStack {
                         Spacer()
                         Text("\(viewModel.message.count)/\(viewModel.limit)")
-                            .font(.system(size: 16, weight: .semibold))
+                            .font(.system(size: 16, weight: .medium))
                             .foregroundColor(LDColor.color3)
                             .padding(.trailing, 14)
                             .padding(.bottom, 10)
