+++ Projects/LastDance/LastDanceClip/Sources/View/ClipRootView.swift
@@ -1,5 +1,5 @@
 //
-//  ContentView.swift
+//  ClipRootView.swift
 //  LastDanceClip
 //
 //  Created by 배현진 on 11/10/25.
